requirements.txt
Flask
hnswlib
pandas
numpy
sentence-transformers
torch
google-generativeai
python-dotenv
gunicorn
faiss-cpu
requests
googletrans
matplotlib
